clean up texture wrapping constants.
